Sorry for one commit :(

+++ GameEngine/Unit.cs
@@ -6,22 +6,22 @@ namespace GameEngine
     {
         public string Type { get; }
 
-        public ulong HitPoints { get; }
+        public int HitPoints { get; }
 
-        public ulong Attack { get; }
+        public int Attack { get; }
 
-        public ulong Defense { get; }
+        public int Defense { get; }
 
-        public (ulong, ulong) Damage { get; }
+        public (int, int) Damage { get; }
 
         public double Initiative { get; }
 
         public Unit(
             string type,
-            ulong hitPoints,
-            ulong attack,
-            ulong defense,
-            (ulong, ulong) damage,
+            int hitPoints,
+            int attack,
+            int defense,
+            (int, int) damage,
             double initiative)
         {
             Type = type ?? throw new ArgumentNullException(nameof(type), "Unit type shoud not be null");
